% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kitten.R
\name{dog}
\alias{dog}
\title{Download a dog}
\usage{
dog(width, height, destfile, ...)
}
\arguments{
\item{width}{Width, in pixels}

\item{height}{Height, in pixels}

\item{destfile}{File to save into.}

\item{...}{Additional parameters passed through to
\code{\link{download.file}}}
}
\value{
The filename saved in (useful when using the default
\code{destfile})
}
\description{
Download a dog picture from placekitten.com
}

