% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitten_function.R
\name{kitten_function}
\alias{kitten_function}
\title{A Kittenlove Function}
\usage{
kitten_function(love = TRUE)
}
\arguments{
\item{love}{Do you love kittens? Defaults to TRUE.}
}
\description{
This function allows you to express your love of kittens.
}
\examples{
kitten_function()
}
\keyword{kittens}
